Error Fix

+++ techpack/audio/4.0/dsp/q6lsm.c
@@ -1100,7 +1100,7 @@ int q6lsm_set_afe_data_format(uint64_t fe_id, uint16_t afe_data_format)
 			lsm_client_afe_data[n].unprocessed_data =
 							afe_data_format;
 			pr_debug("%s: session ID is %d, fe_id is %d\n",
-				 __func__, n, fe_id);
+				 __func__, (int)n, (int)fe_id);
 			return 0;
 		}
 	}
@@ -1133,7 +1133,7 @@ void q6lsm_get_afe_data_format(uint64_t fe_id, uint16_t *afe_data_format)
 			*afe_data_format =
 				lsm_client_afe_data[n].unprocessed_data;
 			pr_debug("%s: session: %d, fe_id: %d, afe data: %s\n",
-				__func__, n, fe_id,
+				__func__, (int)n, (int)fe_id,
 				*afe_data_format ? "unprocessed" : "processed");
 			return;
 		}
